requirements.txt
asyncio
websockets==12.0

google-cloud-speech==2.27.0
google-auth==2.34.0
google-api-core==2.19.2
protobuf==4.25.3

urllib3==2.2.2
